requirements.txt
Django>=5.0.6,<5.2
djangorestframework>=3.15.1
celery>=5.3.4
redis>=5.0.1
psycopg2-binary>=2.9.9; python_version < '3.13'
psycopg[binary]>=3.1.0; python_version >= '3.13'
gunicorn>=21.2.0
whitenoise>=6.6.0
dj-database-url>=2.1.0
python-dotenv>=1.0.0
feedparser>=6.0.11
requests>=2.31.0
django-cors-headers>=4.3.1
pyautogen>=0.4.0  # Python 3.13+ compatible version
pydantic>=2.0.0
ollama>=0.1.0
boto3>=1.34.0
supabase>=2.0.0
